+++ griptape/engines/__init__.py
@@ -10,6 +10,7 @@ from .image.prompt_image_generation_engine import PromptImageGenerationEngine
 from .image.variation_image_generation_engine import VariationImageGenerationEngine
 from .image.inpainting_image_generation_engine import InpaintingImageGenerationEngine
 from .image.outpainting_image_generation_engine import OutpaintingImageGenerationEngine
+from .image_query.image_query_engine import ImageQueryEngine
 
 __all__ = [
     "BaseQueryEngine",
@@ -24,4 +25,5 @@ __all__ = [
     "VariationImageGenerationEngine",
     "InpaintingImageGenerationEngine",
     "OutpaintingImageGenerationEngine",
+    "ImageQueryEngine",
 ]